Refactor databyte.ts Move converters.
